filling out classes and switch

+++ AOC2/project1/project1/ShapeClass.h
@@ -15,7 +15,7 @@
     int numSides;
 }
 
--(NSInteger)getArea;
+-(int)getArea;
 
 -(int)getNumSides;
 -(NSString*)getName;
