+++ AOC2/project1/project1/ShapeClass.m
@@ -20,7 +20,7 @@
     
     return self;
 }
--(NSInteger)getArea
+-(int)getArea
 {
     return 0;
 }
