+++ AOC2/project1/project1/SquareClass.h
@@ -10,5 +10,5 @@
 
 @interface SquareClass : ShapeClass
 
--(NSInteger)getArea;
+-(int)getArea;
 @end
